requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
Django==5.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
idna==3.10
itypes==1.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mysql-connector-python==9.1.0
requests==2.32.3
setuptools==75.1.0
sqlparse==0.5.1
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.3
whitenoise==6.8.2
